Avoid cooling down the extruder in PRINT_START
The current `PRINT_START` macro heats the extruder to 150 degrees, then
to 170, and finally to the target temperature. If the extruder is
already hot, this unnecessarily cools it down and wastes time.

This change only cools down the hotend if it's above the target
temperature and then only to that temperature.

If the extruder was colder than 150 degrees, this change maintains the
current behaviour.